user/app/requirements.txt
#pdf2image==1.17.0
#opencv-python==4.9.0.80
numpy==1.26.4
#pillow==10.3.0
fastapi==0.111.0
easyocr==1.7.1
#pytesseract==0.3.10
#pymupdf
#glob2==0.7
pymysql==1.1.1
uvicorn==0.30.1
sqlalchemy
python-multipart==0.0.7
pillow==10.1.0
elasticsearch==7.15.2
pandas
sentence-transformers
openai-clip
torchvision
jinja2
boto3
s3fs
python-dotenv
python-bidi==0.4.2
openai